authorisasi flash data

+++ app/Http/Controllers/PostController.php
@@ -47,7 +47,10 @@ class PostController extends Controller
         ]);
         $attr['slug'] = \Str::slug(request('title'));
         $attr['category_id'] = request('category');
-        $post = Post::create($attr);
+        //$attr['user_id'] = auth()->id();
+        
+        //$post = Post::create($attr);
+        $post = auth()->user()->posts()->create($attr);
         $post->tags()->attach(request('tags'));
         session()->flash('success','The post was created');
         return redirect('posts');
@@ -80,9 +83,19 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        $post->tags()->detach();
-        $post->delete();
-        session()->flash('success','The post was destroyed');
-        return redirect('posts');
+        // $post->tags()->detach();
+        // $post->delete();
+        // session()->flash('success','The post was destroyed');
+        // return redirect('posts');
+
+        if(auth()->user()->is($post->author)){
+            $post->tags()->detach();
+            $post->delete();
+            session()->flash('success','The post was destroyed');
+            return redirect('posts');
+        } else {
+            session()->flash("error","It wasn't your post");
+            return redirect('posts');
+        }
     }
 }
